Change GelbooruBridge to use Gelbooru API, plus all inheriting Bridges
