+++ bridges/GelbooruBridge.php
@@ -1,35 +1,83 @@
 <?php
-require_once('DanbooruBridge.php');
-
-class GelbooruBridge extends DanbooruBridge {
+class GelbooruBridge extends BridgeAbstract {
 
 	const MAINTAINER = 'mitsukarenai';
 	const NAME = 'Gelbooru';
-	const URI = 'http://gelbooru.com/';
+	const URI = 'https://gelbooru.com/';
 	const DESCRIPTION = 'Returns images from given page';
 
-	const PATHTODATA = '.thumb';
-	const IDATTRIBUTE = 'id';
-	const TAGATTRIBUTE = 'title';
-
-	const PIDBYPAGE = 63;
+	const PARAMETERS = array(
+		'global' => array(
+			'p' => array(
+				'name' => 'page',
+				'defaultValue' => 0,
+				'type' => 'number'
+			),
+			't' => array(
+				'name' => 'tags',
+				'exampleValue' => 'pinup',
+				'title' => 'Tags to search for'
+			),
+			'l' => array(
+				'name' => 'limit',
+				'exampleValue' => 100,
+				'title' => 'How many posts to retrieve (hard limit of 1000)'
+			)
+		),
+		0 => array()
+	);
 
 	protected function getFullURI(){
 		return $this->getURI()
-		. 'index.php?page=post&s=list&pid='
-		. ($this->getInput('p') ? ($this->getInput('p') - 1) * static::PIDBYPAGE : '')
+		. 'index.php?&page=dapi&s=post&q=index&json=1&pid=' . $this->getInput('p')
+		. '&limit=' . $this->getInput('l')
 		. '&tags=' . urlencode($this->getInput('t'));
 	}
 
-	protected function getTags($element){
-		$tags = parent::getTags($element);
-		$tags = explode(' ', $tags);
+	/*
+	This function is superfluous for GelbooruBridge, but useful
+	for Bridges that inherit from it
+	*/
+	protected function buildThumbnailURI($element){
+		return $this->getURI() . 'thumbnails/' . $element->directory
+		. '/thumbnail_' . $element->md5 . '.jpg';
+	}
 
-		// Remove statistics from the tags list (identified by colon)
-		foreach($tags as $key => $tag) {
-			if(strpos($tag, ':') !== false) unset($tags[$key]);
+	protected function getItemFromElement($element){
+		$item = array();
+		$item['uri'] = $this->getURI() . 'index.php?page=post&s=view&id='
+		. $element->id;
+		$item['postid'] = $element->id;
+		$item['author'] = $element->owner;
+		$item['timestamp'] = date('d F Y H:i:s', $element->change);
+		$item['tags'] = $element->tags;
+		$item['title'] = $this->getName() . ' | ' . $item['postid'];
+
+		if (isset($element->preview_url)) {
+			$thumbnailUri = $element->preview_url;
+		} else{
+			$thumbnailUri = $this->buildThumbnailURI($element);
 		}
 
-		return implode(' ', $tags);
+		$item['content'] = '<a href="' . $item['uri'] . '"><img src="'
+		. $thumbnailUri	. '" /></a><br><br><b>Tags:</b> '
+		. $item['tags'] . '<br><br>' . $item['timestamp'];
+
+		return $item;
+	}
+
+	public function collectData(){
+		$content = getContents($this->getFullURI());
+
+		// Most other Gelbooru-based boorus put their content in the root of
+		// the JSON. This check is here for Bridges that inherit from this one
+		$posts = json_decode($content);
+		if (isset($posts->post)) {
+			$posts = $posts->post;
+		}
+
+		foreach($posts as $post) {
+			$this->items[] = $this->getItemFromElement($post);
+		}
 	}
 }
