+++ bridges/MspabooruBridge.php
@@ -5,8 +5,11 @@ class MspabooruBridge extends GelbooruBridge {
 
 	const MAINTAINER = 'mitsukarenai';
 	const NAME = 'Mspabooru';
-	const URI = 'http://mspabooru.com/';
+	const URI = 'https://mspabooru.com/';
 	const DESCRIPTION = 'Returns images from given page';
-	const PIDBYPAGE = 50;
 
+	protected function buildThumbnailURI($element){
+		return $this->getURI() . 'thumbnails/' . $element->directory
+		. '/thumbnail_' . $element->image;
+	}
 }
