+++ bridges/SafebooruBridge.php
@@ -8,5 +8,9 @@ class SafebooruBridge extends GelbooruBridge {
 	const URI = 'https://safebooru.org/';
 	const DESCRIPTION = 'Returns images from given page';
 
-	const PIDBYPAGE = 40;
+	protected function buildThumbnailURI($element){
+		$regex = '/\.\w+$/';
+		return $this->getURI() . 'thumbnails/' . $element->directory
+		. '/thumbnail_' . preg_replace($regex, '.jpg', $element->image);
+	}
 }
