+++ bridges/XbooruBridge.php
@@ -8,5 +8,8 @@ class XbooruBridge extends GelbooruBridge {
 	const URI = 'https://xbooru.com/';
 	const DESCRIPTION = 'Returns images from given page';
 
-	const PIDBYPAGE = 50;
+	protected function buildThumbnailURI($element){
+		return $this->getURI() . 'thumbnails/' . $element->directory
+		. '/thumbnail_' . $element->hash . '.jpg';
+	}
 }
